# api/openapi.yaml
openapi: 3.0.1
info:
  title: Certificate Management Platform API
  version: 1.0.0
  description: |
    Enterprise SSL/TLS Certificate Automation Platform API
    
    Provides endpoints for certificate issuance, inventory management, agent control, authentication, and webhooks.
servers:
  - url: http://localhost:8082/api/v1
    description: Local development
  - url: https://api.cmp.example.com/api/v1
    description: Production
...
paths:
...
  /auth/login:
    post:
      tags:
        - auth
      summary: User login
      description: Authenticate user and receive JWT token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
        '500':
          description: Internal server error
  /auth/me:
    get:
      tags:
        - auth
      summary: Get current user
      description: Get authenticated user information
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /auth/logout:
    post:
      tags:
        - auth
      summary: User logout
      description: Logout current user
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          description: Unauthorized

  # Certificate endpoints
...
  /inventory:
    get:
      tags:
        - inventory
      summary: Get certificate inventory
      description: Discover and list all certificates in the inventory
      operationId: getInventory
      security:
        - BearerAuth: []
      parameters:
        - name: source
          in: query
          schema:
            type: string
        - name: expired
          in: query
          schema:
            type: boolean
        - name: expiring_soon
          in: query
          schema:
            type: integer
            description: Days until expiry
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  certificates:
                    type: array
                    items:
                      $ref: '#/components/schemas/Certificate'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /inventory/expiring:
    get:
      tags:
        - inventory
      summary: Get expiring certificates
      description: Get certificates expiring within specified days
      operationId: getExpiringCertificates
      security:
        - BearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
            description: Number of days to look ahead
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  certificates:
                    type: array
                    items:
                      $ref: '#/components/schemas/Certificate'
                  summary:
                    type: object
                    properties:
                      total:
                        type: integer
                      expiring_7d:
                        type: integer
                      expiring_15d:
                        type: integer
                      expiring_30d:
                        type: integer
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  # Agent endpoints
  /agents:
    get:
      tags:
        - agents
      summary: List agents
      description: Get a list of registered agents
      operationId: listAgents
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /agents/{id}/install:
    post:
      tags:
        - agents
      summary: Install certificate on agent
      description: Request an agent to install a certificate
      operationId: installCertificate
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cert_id
                - path
              properties:
                cert_id:
                  type: string
                path:
                  type: string
                reload_cmd:
                  type: string
      responses:
        '202':
          description: Installation request accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
        '404':
          description: Agent or certificate not found
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  # Events/Webhooks
...
components:
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
        expires_at:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        team:
          type: string
...
    Certificate:
      type: object
      properties:
        id:
          type: string
        fingerprint:
          type: string
        subject:
          type: string
        sans:
          type: array
          items:
            type: string
        issuer:
          type: string
        not_before:
          type: string
          format: date-time
        not_after:
          type: string
          format: date-time
        key_algo:
          type: string
        key_size:
          type: integer
        status:
          type: string
          enum: [active, expired, revoked, pending]
        owner_id:
          type: string
        last_scanned_at:
          type: string
          format: date-time
        source:
          type: string
        days_until_expiry:
          type: integer
...
    Agent:
      type: object
      properties:
        id:
          type: string
        hostname:
          type: string
        ip:
          type: string
        last_checkin:
          type: string
          format: date-time
        status:
          type: string
          enum: [online, offline, error]
        capabilities:
          type: object
          additionalProperties: true
        install_paths:
          type: array
          items:
            type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - BearerAuth: []